back/requirements.txt
asgiref==3.4.1
Django==4.0.1
django-auto-prefetching==0.2.11
django-filter==21.1
djangorestframework==3.13.1
numpy==1.22.0
pandas==1.3.5
python-dateutil==2.8.2
pytz==2021.3
six==1.16.0
sqlparse==0.4.2
PyJWT~=2.3.0
filetype~=1.0.9
drf-yasg==1.20.0